+++ .changeset/little-trains-begin.md
@@ -1,5 +1,5 @@
 ---
-"@trigger.dev/build": minor
+"@trigger.dev/python": minor
 ---
 
 Introduced a new Python extension to enhance the build process. It now allows users to execute Python scripts with improved support and error handling.
