+++ packages/python/CHANGELOG.md
@@ -1,0 +1,1 @@
+# @trigger.dev/python
