+++ packages/python/README.md
@@ -1,0 +1,3 @@
+# Official Python extension for Trigger.dev
+
+View the full documentation [here](https://trigger.dev/docs)
