+++ packages/python/package.json
@@ -1,0 +1,93 @@
+{
+  "name": "@trigger.dev/python",
+  "version": "3.3.16",
+  "description": "Python runtime and build extension for Trigger.dev",
+  "license": "MIT",
+  "publishConfig": {
+    "access": "public"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/triggerdotdev/trigger.dev",
+    "directory": "packages/python"
+  },
+  "type": "module",
+  "files": [
+    "dist"
+  ],
+  "tshy": {
+    "selfLink": false,
+    "main": true,
+    "module": true,
+    "project": "./tsconfig.src.json",
+    "exports": {
+      "./package.json": "./package.json",
+      ".": "./src/index.ts",
+      "./extension": "./src/extension.ts"
+    },
+    "sourceDialects": [
+      "@triggerdotdev/source"
+    ]
+  },
+  "typesVersions": {
+    "*": {
+      "extension": [
+        "dist/commonjs/extension.d.ts"
+      ]
+    }
+  },
+  "scripts": {
+    "clean": "rimraf dist",
+    "build": "tshy && pnpm run update-version",
+    "dev": "tshy --watch",
+    "typecheck": "tsc --noEmit -p tsconfig.src.json",
+    "update-version": "tsx ../../scripts/updateVersion.ts",
+    "check-exports": "attw --pack ."
+  },
+  "dependencies": {
+    "@trigger.dev/build": "workspace:3.3.16",
+    "@trigger.dev/core": "workspace:3.3.16",
+    "@trigger.dev/sdk": "workspace:3.3.16",
+    "tinyexec": "^0.3.2"
+  },
+  "devDependencies": {
+    "@types/node": "20.14.14",
+    "rimraf": "6.0.1",
+    "tshy": "^3.0.2",
+    "typescript": "^5.5.4",
+    "tsx": "4.17.0",
+    "esbuild": "^0.23.0",
+    "@arethetypeswrong/cli": "^0.15.4"
+  },
+  "engines": {
+    "node": ">=18.20.0"
+  },
+  "exports": {
+    "./package.json": "./package.json",
+    ".": {
+      "import": {
+        "@triggerdotdev/source": "./src/index.ts",
+        "types": "./dist/esm/index.d.ts",
+        "default": "./dist/esm/index.js"
+      },
+      "require": {
+        "types": "./dist/commonjs/index.d.ts",
+        "default": "./dist/commonjs/index.js"
+      }
+    },
+    "./extension": {
+      "import": {
+        "@triggerdotdev/source": "./src/extension.ts",
+        "types": "./dist/esm/extension.d.ts",
+        "default": "./dist/esm/extension.js"
+      },
+      "require": {
+        "types": "./dist/commonjs/extension.d.ts",
+        "default": "./dist/commonjs/extension.js"
+      }
+    }
+  },
+  "main": "./dist/commonjs/index.js",
+  "types": "./dist/commonjs/index.d.ts",
+  "module": "./dist/esm/index.js"
+}
