+++ references/v3-catalog/trigger.config.ts
@@ -7,7 +7,7 @@ import { ffmpeg, syncEnvVars } from "@trigger.dev/build/extensions/core";
 import { puppeteer } from "@trigger.dev/build/extensions/puppeteer";
 import { prismaExtension } from "@trigger.dev/build/extensions/prisma";
 import { emitDecoratorMetadata } from "@trigger.dev/build/extensions/typescript";
-import { pythonExtension } from "@trigger.dev/build/extensions/python";
+import { pythonExtension } from "@trigger.dev/python/extension";
 import { defineConfig } from "@trigger.dev/sdk/v3";
 
 export { handleError } from "./src/handleError.js";
